Moved toy data to sccellfie.datasets
