+++ sccellfie/io/tests/test_load_data.py
@@ -10,7 +10,7 @@ from sccellfie.gene_score import compute_gene_scores
 from sccellfie.metabolic_task import compute_mt_score
 from sccellfie.reaction_activity import compute_reaction_activity
 from sccellfie.spatial import create_knn_network
-from sccellfie.tests.toy_inputs import create_random_adata, create_controlled_adata_with_spatial, create_controlled_gpr_dict, create_global_threshold, create_controlled_task_by_rxn
+from sccellfie.datasets.toy_inputs import create_random_adata, create_controlled_adata_with_spatial, create_controlled_gpr_dict, create_global_threshold, create_controlled_task_by_rxn
 
 
 @pytest.mark.parametrize("reactions_filename, mt_filename", [
