+++ sccellfie/tests/test_metabolic_task.py
@@ -5,7 +5,7 @@ from sccellfie.gene_score import compute_gene_scores
 from sccellfie.reaction_activity import compute_reaction_activity
 from sccellfie.metabolic_task import compute_mt_score
 from sccellfie.tests import PCOUNT
-from sccellfie.tests.toy_inputs import create_controlled_adata, create_controlled_gpr_dict, create_global_threshold, create_controlled_task_by_rxn
+from sccellfie.datasets.toy_inputs import create_controlled_adata, create_controlled_gpr_dict, create_global_threshold, create_controlled_task_by_rxn
 
 
 @pytest.mark.parametrize("use_specificity", [True, False])
