+++ sccellfie/tests/test_reaction_activity.py
@@ -5,7 +5,7 @@ import pandas as pd
 from sccellfie.gene_score import compute_gene_scores
 from sccellfie.reaction_activity import compute_reaction_activity
 from sccellfie.tests import PCOUNT
-from sccellfie.tests.toy_inputs import create_controlled_adata, create_controlled_gpr_dict, create_global_threshold
+from sccellfie.datasets.toy_inputs import create_controlled_adata, create_controlled_gpr_dict, create_global_threshold
 
 
 @pytest.mark.parametrize("use_specificity", [True, False])
